리뷰상세 버튼 누르면 모달 나오고 리뷰데이터 출력

+++ src/main/java/com/example/ajouparking/controller/FavoriteController.java
@@ -4,7 +4,6 @@ import com.example.ajouparking.dto.CommonResponseDto;
 import com.example.ajouparking.dto.CustomUserDetails;
 import com.example.ajouparking.entity.User;
 import com.example.ajouparking.service.FavoriteService;
-import com.example.ajouparking.service.LikesService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +21,7 @@ public class FavoriteController {
     private final FavoriteService favoriteService;
 
     @PostMapping("/{parkinglotId}")
-    public ResponseEntity<?> addFavorite(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long parkinglotId){
+    public ResponseEntity<?> addFavorite(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable int parkinglotId){
         User user = customUserDetails.getUser();
         favoriteService.addFavorite(user.getId(),parkinglotId);
 
@@ -31,7 +30,7 @@ public class FavoriteController {
     }
 
     @DeleteMapping("/{parkinglotId}")
-    public ResponseEntity<?> deleteFavorite(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable Long parkinglotId){
+    public ResponseEntity<?> deleteFavorite(@AuthenticationPrincipal CustomUserDetails customUserDetails, @PathVariable int parkinglotId){
         User user = customUserDetails.getUser();
         favoriteService.deleteFavorite(user.getId(), parkinglotId);
         return new ResponseEntity<>(new CommonResponseDto<>("삭제성공",null),HttpStatus.NO_CONTENT);
